requirements.txt
termcolor
coloredlogs
click
pyyaml
six
walrus
